handle seach props in root
