--- React/src/app/pages/game/components/SelectBox/index.ts
@@ -1,1 +1,0 @@
-export * from "./SelectBox";
